requirements.txt
Django==4.0.5
uvicorn==0.17.6
SQLAlchemy==1.4.37
pyodbc==4.0.32
Jinja2==3.1.2
python-dotenv==0.20.0
weasyprint==55.0
python-barcode==0.14.0
qrcode==7.3.1